python/pyproject.toml
[project]
name = "comms-ccf"
version = "0.1.0"
authors = [
    { name = "Kovacsics Robert", email = "kovirobi@gmail.com" }
]
description = "Python tools for the comms-ccf project"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    'cbor2',
    'cobs',
    'cxxfilt',
    'fnv-hash-fast',
    'intervaltree',
    'pyelftools',
    'pyserial-asyncio',
    'squarify',
 ]

[build-system]
requires = ["setuptools"]

[tool.setuptools]
py-modules = [ 'comms_ccf', 'binutils_mapfile' ]

[tool.isort]
profile = "black"
src_paths = [ "comms_ccf", "binutils_mapfile" ]

[project.scripts]
binutils-mapfile = "binutils_mapfile.__main__:main"
comms-ccf = "comms_ccf.__main__:main"
hexdump = "comms_ccf.hexdump:main"

[project.urls]
Homepage = "https://github.com/KoviRobi/comms-ccf"
Issues = "https://github.com/KoviRobi/comms-ccf/issues"
